moved ip_api format options into whatismyip, removed ip_api
